+++ src/Main.hs
@@ -13,8 +13,10 @@ import System.Exit
 main :: IO ()
 main = ssl - do
   _options <- execParser opts
-  r <- runExceptT - runReaderT moeApp _options
-  case r of
-    Left e -> error_ e >> exitFailure
-    Right _ -> pure ()
+  
+  withGateOptions _options - do
+    r <- runExceptT - runReaderT moeApp _options
+    case r of
+      Left e -> error_ e >> exitFailure
+      Right _ -> pure ()