+++ src/Network/MoeSocks/Config.hs
@@ -34,5 +34,6 @@ defaultMoeOptions = MoeOptions
   , _disableSocks5 = False
   , _obfuscation = False
   , _forbidden_IP = ["127.0.0.1", "0.0.0.0", "::1"]
+  , _listMethods = False
   , _params = []
   }